adding user reward call

+++ src/main/java/com/example/rewards/repository/AwardNumberRepository.java
@@ -3,6 +3,6 @@ package com.example.rewards.repository;
 import com.example.rewards.model.AwardNumber;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AwardNumberRepository extends JpaRepository<AwardNumber, Long> {
+public interface AwardNumberRepository extends JpaRepository<AwardNumber, Integer> {
 }
